requirements.txt
appdirs==1.4.3
arrow==0.10.0
bcrypt==3.1.3
cffi==1.9.1
Django==1.8
django-graphos==0.3.41
docutils==0.11
funcsigs==1.0.2
Jinja2==2.9.6
Markdown==2.3.1
MarkupSafe==1.0
matplotlib==1.4.3
mock==2.0.0
nose==1.3.7
numpy==1.10.1
packaging==16.8
pbr==3.1.1
pycparser==2.17
Pygments==2.2.0
pymongo==2.5.1
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2016.10
redis==2.7.4
requests==1.2.0
six==1.10.0
Sphinx==1.2b1
tornado==3.1
